# swagger_openapi_saludvital.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestión Médica - Salud Vital API
  description: |
    API REST para el sistema de gestión médica integral. 
    Permite la gestión de pacientes, médicos, citas médicas y recetas.
    
    ## Usuarios Predefinidos para Pruebas:
    
    ### 👑 **ADMIN**
    - **Email**: `admin@saludvital.com`
    - **Password**: `admin123`
    
    ### 🩺 **MÉDICO** 
    - **Email**: `medico@saludvital.com`
    - **Password**: `medico123`
    
    ### 👤 **PACIENTE**
    - **Email**: `paciente@saludvital.com`
    - **Password**: `paciente123`
    
    ## Flujo de Autenticación:
    1. Hacer login con alguno de los usuarios predefinidos
    2. Copiar el `accessToken` de la respuesta
    3. Usar en header: `Authorization: Bearer {accessToken}`
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: dev@saludvital.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api
    description: Servidor de Desarrollo
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token. Ejemplo 'Bearer {token}'
  schemas:
...
    RecetaRequest:
      type: object
      required:
        - pacienteId
        - items
      properties:
        pacienteId:
          type: integer
          format: int64
          example: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemRecetaRequest'
    ItemRecetaRequest:
      type: object
      required:
        - medicamentoId
        - dosis
        - frecuencia
      properties:
        medicamentoId:
          type: integer
          format: int64
          example: 1
        dosis:
          type: string
          example: "500mg"
        frecuencia:
          type: string
          example: "Cada 8 horas"
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operación exitosa"
        data:
          type: object
...
paths:
...
  /medico/{id}:
    get:
      tags:
        - 🩺 Médicos
      summary: Obtener médico por ID
      description: Obtiene la información de un médico específico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Médico obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Médico no encontrado
...
  /recetas/nueva:
    post:
      tags:
        - 💊 Recetas
      summary: Emitir nueva receta
      description: |
        Crear una nueva receta médica (solo médicos).
        
        **Validaciones:**
        - Medicamentos deben existir en el catálogo
        - No se pueden recetar medicamentos a los que el paciente es alérgico
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecetaRequest'
      responses:
        '200':
          description: Receta creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Error de validación (alergia, medicamento no existe)
        '403':
          description: Sin permisos (solo médicos)
  /recetas/mis-recetas:
    get:
      tags:
        - 💊 Recetas
      summary: Ver mis recetas
      description: Obtiene las recetas del paciente autenticado
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Recetas obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /recetas/medico:
    get:
      tags:
        - 💊 Recetas
      summary: Ver recetas emitidas
      description: Obtiene las recetas emitidas por el médico autenticado
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Recetas obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Sin permisos (solo médicos)
  /recetas/{id}:
    get:
      tags:
        - 💊 Recetas
      summary: Ver receta específica
      description: Obtiene los detalles de una receta específica
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Receta obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...